fieldAlias folder: remove describe clause "field alias"; rename file
